+++ src/index.html
@@ -8,6 +8,11 @@
   </head>
   <body>
     <h1>To do list</h1>
-    <ul id="to-do-list"></ul>
+    <ul id="to-do-list">
+      <span class="ul"></span>
+      <li class="btn-clear">
+        <button type="button">Clear all completed</button>
+      </li>
+    </ul>
   </body>
 </html>
