+++ src/index.js
@@ -4,14 +4,13 @@ import {
   addTask,
   deleteTaks,
   updateTaks,
-  changeStatus,
   deleteAllTaks,
 } from './modules/functions.js';
+import changeStatus from './modules/statusUpdate.js';
 
-const ul = document.querySelector('ul');
+const ul = document.querySelector('.ul');
 loadTasks(localStorage.getItem('lists'), ul);
 const addInput = document.querySelector('.add-input form');
-const clearAll = document.querySelector('.btn-clear button');
 addInput.addEventListener('submit', (event) => {
   event.preventDefault();
   addTask(document.querySelector('#new-item'), ul);
@@ -22,6 +21,7 @@ addInput.addEventListener('submit', (event) => {
 deleteTaks(ul);
 updateTaks();
 changeStatus();
+const clearAll = document.querySelector('.btn-clear button');
 clearAll.addEventListener('click', () => {
   deleteAllTaks(ul);
 });
